Initial stab at generic database credentials class.
There's no need for separate D::P::Auth::RBAC::Credentials::* classes for each
DB engine.

This is still rough round the edges and needs some tweaking, tidying and
testing.

It's a step in the right direction though, I think.

I've made the use of roles configurable, so if someone wants to use this for a
simple app with just a users table, where all users have the same simple role,
then that's fine.

Roles are stored in normalised form in a roles table, and users are assigned
roles with a user_roles table.

Table and field names can be changed via config, which may  be useful for
someone trying to use this plugin with an existing database (say, basing their
login off some forum software's user DB, or something).

A lot more to come, but getting this out there for eyeballing.